Updating to Blacklight 4.8.0
This resolves an apparent exception in upgrading to Rails 4.2.x.
This moves us forward in UI testing, hopefully the semantic versioning
of Blacklight is trustworthy.

```console
ActionView::Template::Error (undefined method `to_model' for
\#<ActiveSupport::HashWithIndifferentAccess:0x007ff9b2a18700>):
``

Also, as a bonus, allowing for a parameter to run rails server without
an assumed asset precompile.