Fix mysql user/pass
This fixes the connectivity between rails container and mysql. Also adds
x bit to the wait-for-it script. Nginx was failing to wait for the rails
service. Also added a .env that can be used for defaults in development.
Removed the docker-compose-rails.yml since we've merged this into the
base compose.