Fixing "Delete this and replace with" behavior
Prior to this commit, and I suspect due to a Rails upgrade, the
functionality of "Delete this and replace with" behavior for
Classification and Country did not work.

In debugging, it was no longer adequate to declare an `.attr_accessor`
for the virtual attribute used to trigger the delete/update logic
branch.

The `.attr_accessible` declaration is required to expose, for
`#update_attributes`, the fieldused for triggering the delete/update
behavior