Ensuring the "Delete and Replace" feature works
Prior to this commit, the Country, Language, and Classification models
exposed a "Delete and Replace" feature. In Rails 4.2 (and perhaps
before work on Okta-fication began) these behaviors did not work.

The Country, Language, or Classification would be deleted. However,
due to caching of the relationships, the database records would have
the correct "Replaced" entry (except for Country, more on that later).
However, the SOLR document would have the "Deleted" entry. This is
based on timing (First we update the related facsimile, which would
run a SOLR commit, then we run the after delete cleanup method, which
would use the cached relationship to run solr updates).

This change ensures that we void the cache before we attempt to run
the update.

And why was Country / Facsimile a different case? In using
`#country_id` instead of `#country`, Rails updated the ID but not the
`#country` object of the Facsimile. When save was run on the
Facsimile, Rails favored the `#country` value instead of the
`#country_id`.

In other words, setting `Facsimile#country_id` does not update
`Facsimile#country`, but setting `Facsimile.country` does update
`Facsimile#country_id`.